+++ ChangeLog.md
@@ -1,3 +1,9 @@
+### 0.9.30.8:
+* Fix crashes on Android < 4.2
+
+### 0.9.30.7:
+* Fix crashes on Android < 4.2
+
 ### 0.9.30.5:
 * Make app compatible with Marshmallow